% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution par jour de la semaine}
\usage{
calcul_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets}
}
\value{
Un data.frame avec une ligne pour chaque jour
}
\description{
Cette fonction calcule le nombre total de trajets par jour
}
