% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucle}
\usage{
compter_nombre_boucle(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets}
}
\value{
Un data.frame avec deux colonnes
}
\description{
Cette fonction compte le nombre total de boucle
}
