% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre total de trajets}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets}
}
\value{
Un entier le nombre total de trajets
}
\description{
Cette fonction compte le nombre total de trajets dans un data.frame
}
