% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{Jeu de données des trajets en vélo à Nantes.}
\format{
Un data.frame avec les colonnes de comptage
}
\source{
\url{https://data.nantesmetropole.fr/}
}
\usage{
df_velo
}
\description{
Données issue de l'open data de nantes métropole
}
\keyword{datasets}
