% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data frame contenant les trajets}
}
\value{
Un data frame sans les observations anormales
}
\description{
Cette fonction retire les observations considérées comme anormales
}
