% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_trajet.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets selon les boucles selectionnees}
\usage{
filtrer_trajet(trajet, boucle)
}
\arguments{
\item{trajet}{data.frame. Le jeu de donnees contenant les comptages.}

\item{boucle}{character. Un vecteur de numeros de boucle a conserver.}
}
\value{
Un data.frame filtre contenant uniquement les boucles selectionnees.
}
\description{
Cette fonction filtre un data.frame contenant des comptages velo
pour ne conserver que les lignes correspondant aux boucles indiquees.
}
\examples{
filtrer_trajet(trajet = df_velo, boucle = c("880", "881"))

}
