% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Faire une graphique de la distribution par jour de la semaine}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets}
}
\value{
Un graphique représentant la distibrution par jour
}
\description{
Cette fonction calcule le nombre total de trajets par jour
}
