% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet max}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets}
}
\value{
Un data.frame avec une ligne
}
\description{
Cette fonction trouve la boucle ayant enregistrée le plus grand nombre de passages
}
